# abacusflow-portal/abacusflow-portal-web/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: abacusflow api
  version: 0.0.1
  description: 小算盘服务接口文档
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /platform/tenants:
    get:
      operationId: listPlatformTenants
      tags:
        - tenant
      summary: List every tenant in the platform control plane
      responses:
        "200":
          description: Platform tenant directory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformTenant"
    post:
      operationId: provisionTenant
      tags:
        - tenant
      summary: Provision a pending tenant and its initial administrator invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantInput"
      responses:
        "201":
          description: Tenant provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantProvisioning"
  /platform/tenants/{tenantId}:
    put:
      operationId: updatePlatformTenant
      tags:
        - tenant
      summary: Update any tenant control-plane profile
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantInput"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformTenant"
  /platform/tenants/{tenantId}/initial-invitation/reissue:
    post:
      operationId: reissueInitialTenantInvitation
      tags:
        - tenant
      summary: Cancel and reissue a pending tenant's initial administrator invitation
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReissueInitialInvitationInput"
      responses:
        "200":
          description: One-time invitation delivery result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantInvitation"
...
  /platform/permissions/{permissionId}:
    put:
      operationId: updatePermission
      tags:
        - permission
      summary: Update a permission
      description: 更新权限的显示名称和描述，需要 platform:permission:manage 权限
      parameters:
        - name: permissionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePermissionInput"
      responses:
        "200":
          description: Permission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Permission"
        "403":
          description: 无权限
        "404":
          description: 权限不存在
  /platform/roles:
    get:
      operationId: listPlatformRoles
      tags:
        - platform-role
      summary: List global platform roles
      responses:
        "200":
          description: Platform roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformRole"
    post:
      operationId: createPlatformRole
      tags:
        - platform-role
      summary: Create a global platform role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformRoleInput"
      responses:
        "201":
          description: Platform role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformRole"
  /platform/roles/{roleId}:
    put:
      operationId: updatePlatformRole
      tags:
        - platform-role
      summary: Update a global platform role
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformRoleInput"
      responses:
        "200":
          description: Platform role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformRole"
    delete:
      operationId: deletePlatformRole
      tags:
        - platform-role
      summary: Delete a global platform role
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Platform role deleted
  /platform/roles/{roleId}/users/{userId}:
    post:
      operationId: assignPlatformRole
      tags:
        - platform-role
      summary: Assign a global platform role to a user
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "201":
          description: Platform role assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformRoleAssignment"
    delete:
      operationId: removePlatformRole
      tags:
        - platform-role
      summary: Remove a global platform role from a user
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Platform role removed
  /platform/roles/{roleId}/users:
    get:
      operationId: listPlatformRoleAssignments
      tags:
        - platform-role
      summary: List users assigned to a global platform role
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Platform role assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformRoleAssignment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateTenantInput:
      type: object
      required:
        - name
        - initialAdministratorEmail
      properties:
        name:
          type: string
          description: 租户名称
          maxLength: 50
        displayName:
          type: string
          nullable: true
          description: 租户显示名称
        initialAdministratorEmail:
          type: string
          format: email
          description: 首位租户管理员邮箱
    ReissueInitialInvitationInput:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    PlatformTenant:
      type: object
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        displayName:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING_ACTIVATION, ACTIVE, SUSPENDED, DEPROVISIONED]
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
    TenantProvisioning:
      type: object
      required:
        - tenant
        - initialInvitation
      properties:
        tenant:
          $ref: "#/components/schemas/PlatformTenant"
        initialInvitation:
          $ref: "#/components/schemas/TenantInvitation"
...
    UpdateTenantInput:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
          description: 租户显示名称
          maxLength: 200
...
    TenantInvitation:
      type: object
      required:
        - id
        - tenantId
        - tenantName
        - email
        - roleIds
        - roleNames
        - status
        - expiresAt
        - createdAt
        - initialAdministrator
      properties:
        id:
          type: integer
          format: int64
          description: 邀请 ID
        tenantId:
          type: integer
          format: int64
          description: 目标租户 ID
        tenantName:
          type: string
          description: 目标租户名称
        email:
          type: string
          description: 被邀请人邮箱
        roleIds:
          type: array
          items:
            type: integer
            format: int64
          description: 邀请时指定的角色 ID 列表
        roleNames:
          type: array
          items:
            type: string
          description: 角色名称列表
        invitedByUserId:
          type: integer
          format: int64
          nullable: true
          description: 发起邀请的用户 ID
        token:
          type: string
          nullable: true
          description: 邀请唯一 token；仅在创建或重发响应中返回一次，列表和接受响应不会返回
        status:
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - CANCELLED
          description: 邀请状态
        expiresAt:
          type: integer
          format: int64
          description: 过期时间戳（毫秒）
        acceptedAt:
          type: integer
          format: int64
          nullable: true
          description: 接受时间戳（毫秒）
        createdAt:
          type: integer
          format: int64
          description: 创建时间戳（毫秒）
        initialAdministrator:
          type: boolean
          description: 是否为待激活租户的首位管理员邀请
...
    Permission:
      type: object
      required:
        - id
        - name
        - label
        - description
        - scope
      properties:
        id:
          type: integer
          format: int64
          description: 权限 ID
        name:
          type: string
          description: 权限名称（如 business:product:read）
        label:
          type: string
          description: 显示名称
        description:
          type: string
          description: 权限描述
        scope:
          type: string
          enum: [PLATFORM, TENANT, BUSINESS]
    PlatformRole:
      type: object
      required: [id, name, label, permissionNames]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        label:
          type: string
        permissionNames:
          type: array
          items:
            type: string
    PlatformRoleInput:
      type: object
      required: [name, permissionIds]
      properties:
        name:
          type: string
        label:
          type: string
          nullable: true
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
    PlatformRoleAssignment:
      type: object
      required: [userId, userName, roleId, roleName]
      properties:
        userId:
          type: integer
          format: int64
        userName:
          type: string
        roleId:
          type: integer
          format: int64
        roleName:
          type: string
    UpdatePermissionInput:
      type: object
      properties:
        label:
          type: string
          nullable: true
          description: 显示名称
          maxLength: 255
        description:
          type: string
          nullable: true
          description: 权限描述
          maxLength: 500
...